# Optimize constant label pair adding with relabel.Replace