--- a/model/relabel/relabel.go
+++ b/model/relabel/relabel.go
@@ -267,6 +267,17 @@ func relabel(cfg *Config, lb *labels.Builder) (keep bool) {
 			return false
 		}
 	case Replace:
+		// Fast path to add or delete label pair.
+		if val == "" && cfg.Regex == DefaultRelabelConfig.Regex &&
+			!varInRegexTemplate(cfg.TargetLabel) && !varInRegexTemplate(cfg.Replacement) {
+			if !model.LabelName(cfg.TargetLabel).IsValid() || cfg.Replacement == "" {
+				lb.Del(cfg.TargetLabel)
+			} else {
+				lb.Set(cfg.TargetLabel, cfg.Replacement)
+			}
+			break
+		}
+
 		indexes := cfg.Regex.FindStringSubmatchIndex(val)
 		// If there is no match no replacement must take place.
 		if indexes == nil {
@@ -315,4 +326,8 @@ func relabel(cfg *Config, lb *labels.Builder) (keep bool) {
 	}
 
 	return true
+}
+
+func varInRegexTemplate(template string) bool {
+	return strings.Contains(template, "$")
 }