--- a/model/relabel/relabel_test.go
+++ b/model/relabel/relabel_test.go
@@ -15,6 +15,7 @@ package relabel
 
 import (
 	"fmt"
+	"sort"
 	"testing"
 
 	"github.com/prometheus/common/model"
@@ -848,5 +849,64 @@ func BenchmarkRelabel(b *testing.B) {
 				_, _ = Process(tt.lbls, tt.cfgs...)
 			}
 		})
+	}
+}
+
+func BenchmarkRelabel_ReplaceAddLabel(b *testing.B) {
+	cfgs := []*Config{}
+	for k, v := range map[string]string{
+		"wwwwww":       "wwwwww",
+		"xxxxxxxxx":    "xxxxxxxxx",
+		"yyyyyyyyyyyy": "yyyyyyyyyyyy",
+		"${0}":         "dropped",
+		"dropped":      "${0}",
+	} {
+		cfgs = append(cfgs, &Config{
+			Action:      DefaultRelabelConfig.Action,
+			Separator:   DefaultRelabelConfig.Separator,
+			Regex:       DefaultRelabelConfig.Regex,
+			TargetLabel: k,
+			Replacement: v,
+		})
+	}
+	expectLset := labels.Labels{
+		labels.Label{Name: "abcdefg01", Value: "hijklmn1"},
+		labels.Label{Name: "abcdefg02", Value: "hijklmn2"},
+		labels.Label{Name: "abcdefg03", Value: "hijklmn3"},
+		labels.Label{Name: "abcdefg04", Value: "hijklmn4"},
+		labels.Label{Name: "abcdefg05", Value: "hijklmn5"},
+		labels.Label{Name: "abcdefg06", Value: "hijklmn6"},
+		labels.Label{Name: "abcdefg07", Value: "hijklmn7"},
+		labels.Label{Name: "abcdefg08", Value: "hijklmn8"},
+		labels.Label{Name: "abcdefg09", Value: "hijklmn9"},
+		labels.Label{Name: "abcdefg10", Value: "hijklmn10"},
+		labels.Label{Name: "abcdefg11", Value: "hijklmn11"},
+		labels.Label{Name: "abcdefg12", Value: "hijklmn12"},
+		labels.Label{Name: "abcdefg13", Value: "hijklmn13"},
+		labels.Label{Name: "wwwwww", Value: "wwwwww"},
+		labels.Label{Name: "xxxxxxxxx", Value: "xxxxxxxxx"},
+		labels.Label{Name: "yyyyyyyyyyyy", Value: "yyyyyyyyyyyy"},
+	}
+	sort.Sort(expectLset)
+
+	for i := 0; i < b.N; i++ {
+		lset := labels.Labels{
+			labels.Label{Name: "abcdefg01", Value: "hijklmn1"},
+			labels.Label{Name: "abcdefg02", Value: "hijklmn2"},
+			labels.Label{Name: "abcdefg03", Value: "hijklmn3"},
+			labels.Label{Name: "abcdefg04", Value: "hijklmn4"},
+			labels.Label{Name: "abcdefg05", Value: "hijklmn5"},
+			labels.Label{Name: "abcdefg06", Value: "hijklmn6"},
+			labels.Label{Name: "abcdefg07", Value: "hijklmn7"},
+			labels.Label{Name: "abcdefg08", Value: "hijklmn8"},
+			labels.Label{Name: "abcdefg09", Value: "hijklmn9"},
+			labels.Label{Name: "abcdefg10", Value: "hijklmn10"},
+			labels.Label{Name: "abcdefg11", Value: "hijklmn11"},
+			labels.Label{Name: "abcdefg12", Value: "hijklmn12"},
+			labels.Label{Name: "abcdefg13", Value: "hijklmn13"},
+		}
+		actual, _ := Process(lset, cfgs...)
+		var _ = actual
+		// require.Equal(b, actual, expectLset)
 	}
 }