<!--
    Don't forget!

    - Please sign CNCF's Developer Certificate of Origin and sign-off your commits by adding the -s / --sign-off flag to `git commit`. See https://github.com/apps/dco for more information.

    - If the PR adds or changes a behaviour or fixes a bug of an exported API it would need a unit/e2e test.

    - Where possible use only exported APIs for tests to simplify the review and make it as close as possible to an actual library usage.

    - No tests are needed for internal implementation changes.

    - Performance improvements would need a benchmark test to prove it.

    - All exposed objects should have a comment.

    - All comments should start with a capital letter and end with a full stop.
 -->
```
name                       old time/op    new time/op    delta
Relabel/example-8            2.30µs ±69%    1.49µs ±26%  -35.20%  (p=0.000 n=20+20)
Relabel/kubernetes-8         6.02µs ± 5%    5.87µs ± 4%   -2.39%  (p=0.000 n=18+20)
Relabel_ReplaceAddLabel-8     973ns ±16%     546ns ±24%  -43.91%  (p=0.000 n=20+20)

name                       old alloc/op   new alloc/op   delta
Relabel/example-8              633B ± 4%      574B ± 4%   -9.24%  (p=0.000 n=20+20)
Relabel/kubernetes-8         1.46kB ± 1%    1.38kB ± 2%   -5.37%  (p=0.000 n=20+20)
Relabel_ReplaceAddLabel-8    1.81kB ± 0%    1.55kB ± 0%  -14.22%  (p=0.000 n=20+20)

name                       old allocs/op  new allocs/op  delta
Relabel/example-8              17.0 ± 0%      12.0 ± 0%  -29.41%  (p=0.000 n=20+20)
Relabel/kubernetes-8           36.0 ± 0%      31.0 ± 0%  -13.89%  (p=0.000 n=20+20)
Relabel_ReplaceAddLabel-8      19.0 ± 0%       4.0 ± 0%  -78.95%  (p=0.000 n=20+20)
```