@bboreham 2023-03-24
I'm not sure I understand.  What is the significance of looking for `$` ?