@roidelapluie 2023-03-24
This is not correct as the default relabel config does not replace label values. It only replaces label values that do not have a new line in it.