@damnever 2023-03-25
This PR is optimizing the cases like this:
```yaml
- action: replace
  target:  contains_no_variables
  replacement:  contains_no_variables
  separator: ;    // the default
  regex: (.*)     // the default
```
this add an new label pair: `contains_no_variables=contains_no_variables`.

If the regex matches all and there is no variables (the symbol `$`) in the templates (the target and replacement), then we can skip `Regex.FindStringSubmatchIndex` and `Regex.ExpandString`.