@damnever 2023-03-25
> What is the significance of looking for `$` ?

I think you might be right, there is no need to look for `$`.