@bboreham 2023-03-25
> does not contain new lines 

(Might be confusing: this is the result of Prometheus evaluating regexps in multi-line mode. Which is arguably a bug, but not one we can fix without a feature flag or major version bump.)