@roidelapluie 2023-03-25
Yes, this is something we plan to fix in Prometheus 3.