@damnever 2023-03-29
@roidelapluie [this commit check the new line in the target](https://github.com/prometheus/prometheus/pull/12180/commits/3d981c8e2cedf19f1339b626252d9eed2453a1c5)