@damnever 2023-06-28
@roidelapluie I believe that after a successful replacement, the invalid label will be dropped https://github.com/prometheus/prometheus/blob/fe639b545fa8564cf3ea57a7a0b815e432ab7aed/model/relabel/relabel.go#L263-L267