@roidelapluie 2023-12-05
We have looked at this pull request during our bug scrub.

I have added this to the 3.0 milestone where we will fix the default regex, making this pull request simpler and more valuable.


Thank you for your contribution.