@bboreham 2023-12-18
> This PR is optimizing the cases like this:

> 
```
- action: replace
  target:  contains_no_variables
  replacement:  contains_no_variables
  separator: ;    // the default
  regex: (.*)     // the default
```

If you add one more condition:
```
source_labels: // empty (the default)
```

then the whole discussion about newlines goes away. (It was already moot to pass `if val == ""` in your code).

Also `lb.Set` already does `Del` if the replacement is blank so you don't need to check that.
In #13308 I propose a check that `TargetLabel` is valid, in the case it does not contain `$`.

All the above lets the optimization collapse to:
```
		// Fast path to add static label pair.
		if val == "" && cfg.Regex == DefaultRelabelConfig.Regex &&
			!varInRegexTemplate(cfg.TargetLabel) && !varInRegexTemplate(cfg.Replacement) {
			lb.Set(cfg.TargetLabel, cfg.Replacement)
			break
		}
```

I don't think we need a completely new Benchmark function; it should be possible to do your focus as a new case within the existing `BenchmarkRelabel`.