@damnever 2023-12-19
@bboreham great explanation, I think I was somewhat confused before.

After some experiments, I believe the SourceLabels don't really matter either, since the default regex is `(.*)`. Therefore, it seems okay to remove the `val == ""` check as well.