@bboreham 2023-12-19
Please don't do that; we will be back into the newline discussion.
`val == ""` guarantees we don't have to talk about newlines.
`(.*)` does not match `"\n"`, for instance.