@bboreham 2024-07-30
My position is this PR does not need to talk about newlines, therefore let's not.