@bboreham 2024-09-24
BTW this should not have been flagged as a 3.0 change; it should be invisible to users and a pure optimisation.